removed old leaderboard api route
