--- src/app/leaderboard/mockusers.json
@@ -1,130 +1,0 @@
-{
-    "users": [
-      {
-        "id": "1",
-        "name": "Simon Dao",
-        "hackeroons": 1000,
-        "history": [["workshop1", "workshop", 10], ["acm hat", "purchase", -10], ["workshop2", "workshop", 10]]
-      },
-      {
-        "id": "2",
-        "name": "Alex Lee",
-        "hackeroons": 800,
-        "history": [["workshop1", "workshop", 15], ["acm hoodie", "purchase", -20], ["workshop3", "workshop", 20]]
-      },
-      {
-        "id": "3",
-        "name": "Rachel Kim",
-        "hackeroons": 950,
-        "history": [["workshop1", "workshop", 20], ["acm sticker", "purchase", -5], ["workshop2", "workshop", 15]]
-      },
-      {
-        "id": "4",
-        "name": "Michael Smith",
-        "hackeroons": 1200,
-        "history": [["workshop3", "workshop", 25], ["acm mug", "purchase", -10], ["workshop4", "workshop", 30]]
-      },
-      {
-        "id": "5",
-        "name": "Jessica Brown",
-        "hackeroons": 1100,
-        "history": [["workshop2", "workshop", 10], ["acm beanie", "purchase", -15], ["workshop3", "workshop", 20]]
-      },
-      {
-        "id": "6",
-        "name": "David Wong",
-        "hackeroons": 950,
-        "history": [["workshop1", "workshop", 10], ["acm hoodie", "purchase", -20], ["workshop2", "workshop", 15]]
-      },
-      {
-        "id": "7",
-        "name": "Sarah Patel",
-        "hackeroons": 870,
-        "history": [["workshop2", "workshop", 12], ["acm mug", "purchase", -10], ["workshop3", "workshop", 18]]
-      },
-      {
-        "id": "8",
-        "name": "John Clark",
-        "hackeroons": 1020,
-        "history": [["workshop1", "workshop", 20], ["acm beanie", "purchase", -15], ["workshop3", "workshop", 25]]
-      },
-      {
-        "id": "9",
-        "name": "Emma Lee",
-        "hackeroons": 950,
-        "history": [["workshop1", "workshop", 15], ["acm sticker", "purchase", -5], ["workshop4", "workshop", 30]]
-      },
-      {
-        "id": "10",
-        "name": "James Wilson",
-        "hackeroons": 980,
-        "history": [["workshop2", "workshop", 18], ["acm hoodie", "purchase", -20], ["workshop3", "workshop", 22]]
-      },
-      {
-        "id": "11",
-        "name": "Sophia Harris",
-        "hackeroons": 1150,
-        "history": [["workshop3", "workshop", 25], ["acm hat", "purchase", -10], ["workshop2", "workshop", 20]]
-      },
-      {
-        "id": "12",
-        "name": "Daniel Lewis",
-        "hackeroons": 1050,
-        "history": [["workshop1", "workshop", 10], ["acm sticker", "purchase", -5], ["workshop3", "workshop", 30]]
-      },
-      {
-        "id": "13",
-        "name": "Olivia Martinez",
-        "hackeroons": 1080,
-        "history": [["workshop1", "workshop", 12], ["acm mug", "purchase", -10], ["workshop4", "workshop", 25]]
-      },
-      {
-        "id": "14",
-        "name": "Ethan Thomas",
-        "hackeroons": 940,
-        "history": [["workshop2", "workshop", 15], ["acm hoodie", "purchase", -20], ["workshop3", "workshop", 22]]
-      },
-      {
-        "id": "15",
-        "name": "Ava White",
-        "hackeroons": 1100,
-        "history": [["workshop1", "workshop", 10], ["acm beanie", "purchase", -15], ["workshop2", "workshop", 30]]
-      },
-      {
-        "id": "16",
-        "name": "Benjamin Anderson",
-        "hackeroons": 1200,
-        "history": [["workshop3", "workshop", 25], ["acm sticker", "purchase", -5], ["workshop2", "workshop", 30]]
-      },
-      {
-        "id": "17",
-        "name": "Mia Robinson",
-        "hackeroons": 1000,
-        "history": [["workshop1", "workshop", 20], ["acm mug", "purchase", -10], ["workshop3", "workshop", 25]]
-      },
-      {
-        "id": "18",
-        "name": "Liam Young",
-        "hackeroons": 1070,
-        "history": [["workshop2", "workshop", 12], ["acm hoodie", "purchase", -20], ["workshop4", "workshop", 22]]
-      },
-      {
-        "id": "19",
-        "name": "Zoe King",
-        "hackeroons": 1120,
-        "history": [["workshop3", "workshop", 30], ["acm hat", "purchase", -10], ["workshop1", "workshop", 15]]
-      },
-      {
-        "id": "20",
-        "name": "Lucas Scott",
-        "hackeroons": 950,
-        "history": [["workshop2", "workshop", 18], ["acm sticker", "purchase", -5], ["workshop3", "workshop", 25]]
-      },
-      {
-        "id": "21",
-        "name": "Charlotte Allen",
-        "hackeroons": 1050,
-        "history": [["workshop3", "workshop", 20], ["acm hoodie", "purchase", -20], ["workshop4", "workshop", 30]]
-      }
-    ]
-  }
